requirements.txt
firebase_admin==6.4.0
Flask==3.1.1
Flask_Cors==4.0.0
protobuf==6.31.1
pyparsing==3.2.3
python-dotenv==1.1.1
Requests==2.32.4
